Extract the JobDispatcher from ServiceRegistryImpl
This commit also reworks how the JobDispatcher gets enabled, and defaults to *disabled* to prevent multiple dispatchers
from running by default.  This will necessitate a change in the node profiles in a subsequent commit.
