+++ etc/org.opencastproject.serviceregistry.impl.JobDispatcher.cfg
@@ -1,0 +1,13 @@
+# Configuration for the job dispatcher
+
+# Enables dispatching from this node.  It is very important that only a single node be dispatching at once.
+# Usually you want to enable this on your admin node, but nowhere else.
+# Default: false
+# dispatch.enabled=false
+
+# The interval in seconds between two rounds of dispatching in the service registry. A minimum value of 1s is enforced
+# due to performance reasons. Set to 0 to disable dispatching from this service registry.
+# Service registry dispatching is automatically set to 0 on everything but admin or allinone nodes and should usually
+# not be activated on these nodes to avoid concurrency problems.
+# Default: 2
+#dispatch.interval=2
