+++ etc/org.opencastproject.serviceregistry.impl.ServiceRegistryJpaImpl.cfg
@@ -8,13 +8,6 @@
 # Default: None
 #no.error.state.service.types=
 
-# The interval in seconds between two rounds of dispatching in the service registry. A minimum value of 1s is enforced
-# due to performance reasons. Set to 0 to disable dispatching from this service registry.
-# Service registry dispatching is automatically set to 0 on everything but admin or allinone nodes and should usually
-# not be activated on these nodes to avoid concurrency problems.
-# Default: 2
-#dispatch.interval=2
-
 # The interval in seconds between checking if the hosts in the service registry hosts are still alive.
 # Set to 0 to disable checking if hosts are still alive and able to be dispatched to.
 # Default: 60
